Change to return removed_count rather than ErrorCode

+++ mooncake-integration/store/store_py.h
@@ -127,7 +127,7 @@ class DistributedObjectStore {
 
     int remove(const std::string &key);
 
-    int removeAll();
+    long removeAll();
 
     int tearDownAll();
